Fixed the s_addr in InAddr structure. This was raising errors when trying to retrieve addresses with inet_ntop

+++ structures/ifaddrs.py
@@ -51,7 +51,7 @@ class InAddr(ctypes.Structure):
     """
 
     _fields_ = [
-        ('s_addr', ctypes.c_ulong),   # IPV4 Load with inet_aton()
+        ('s_addr', ctypes.c_uint32),   # IPV4 Load with inet_aton()
     ]
